+++ @xen-orchestra/web-core/lib/components/ui/accordion-list/UiAccordionList.vue
